refactor: ProductItem projection interface -> dto 변경 (#134)
