+++ src/main/java/com/example/i_commerce/domain/product/application/dto/ProductItemInfo.java
@@ -1,8 +1,9 @@
-package com.example.i_commerce.domain.cart.infrastructure;
+package com.example.i_commerce.domain.product.application.dto;
 
 
 import com.example.i_commerce.domain.cart.entity.CartItem;
-import com.example.i_commerce.domain.product.facade.dto.ProductItemInfoResponse;
+import com.example.i_commerce.domain.product.entity.enums.ProductItemStatus;
+import com.example.i_commerce.domain.product.repository.projection.ProductItemInfoProjection;
 import lombok.Builder;
 
 @Builder
@@ -16,15 +17,15 @@ public record ProductItemInfo(
     boolean isOnSale
 ) {
 
-    public static ProductItemInfo from(ProductItemInfoResponse response) {
+    public static ProductItemInfo from(ProductItemInfoProjection projection) {
         return ProductItemInfo.builder()
-            .productItemId(response.productItemId())
-            .productName(response.productName())
-            .storeId(response.storeId())
-            .price(response.price())
-            .displayOptionName(response.displayOptionName())
-            .stockQuantity(response.stockQuantity())
-            .isOnSale(response.isOnSale())
+            .productItemId(projection.productItemId())
+            .productName(projection.productName())
+            .storeId(projection.storeId())
+            .price(projection.price())
+            .displayOptionName(projection.displayOptionName())
+            .stockQuantity(projection.stockQuantity())
+            .isOnSale(projection.productItemStatus() == ProductItemStatus.ON_SALE)
             .build();
     }
