+++ src/main/java/com/example/i_commerce/domain/product/repository/projection/ProductItemInfoProjection.java
@@ -1,13 +1,15 @@
 package com.example.i_commerce.domain.product.repository.projection;
 
-import com.example.i_commerce.domain.product.entity.ProductItemStatus;
+import com.example.i_commerce.domain.product.entity.enums.ProductItemStatus;
 
-public interface ProductItemInfoProjection {
-    Long getProductItemId();
-    String getProductName();
-    Long getStoreId();
-    Integer getPrice();
-    String getDisplayOptionName();
-    Integer getStockQuantity();
-    ProductItemStatus getStatus();
+public record ProductItemInfoProjection(
+    Long productItemId,
+    String productName,
+    Long storeId,
+    Integer price,
+    String displayOptionName,
+    Integer stockQuantity,
+    ProductItemStatus productItemStatus
+) {
+
 }
